pyproject.toml
[tool.poetry]
name = "wendah-simple"
version = "0.1.0"
description = ""
authors = ["AiWaldoh <kevin@gptcoder.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
cohere = "^5.5.8"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
playwright = "^1.44.0"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
